+++ src/main/java/io/github/astrapi69/checksum/ByteArrayChecksumExtensions.java
@@ -26,7 +26,7 @@ import java.util.zip.Adler32;
 import java.util.zip.CRC32;
 import java.util.zip.Checksum;
 
-import io.github.astrapi69.crypto.algorithm.Algorithm;
+import io.github.astrapi69.crypt.api.algorithm.Algorithm;
 
 /**
  * The class {@link ByteArrayChecksumExtensions} is a utility class for computing checksum from byte