+++ src/main/java/module-info.java
@@ -19,5 +19,5 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 module checksum.up {
-	requires crypt.api.main;
+	requires crypt.api;
 }